requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-ckeditor==6.5.0
django-js-asset==2.0.0
Pillow==9.2.0
sqlparse==0.4.2
